fixed cmake
